### 2025-02-06 693591b
Shorten paper and add DOI to paper.bib
