+++ paper.md
@@ -1,11 +1,6 @@
 ---
-title: "Syclops: A Modular Pipeline for Procedural Generation of Synthetic Data"
-tags:
-  - Python
-  - synthetic data
-  - procedural generation
-  - computer vision
-  - Blender
+title: "Syclops: A Modular Pipeline for Procedural Generation of Synthetic Data"  
+tags: [Python, synthetic data, procedural generation, computer vision, Blender]  
 authors:
   - name: Anton Elmiger
     corresponding: true
@@ -24,29 +19,30 @@ affiliations:
     index: 2
 date: 22 January 2025
 bibliography: paper.bib
+
 ---
 
-# Summary
+## Summary
 
-Syclops is an open-source, modular pipeline designed for generating large-scale, photorealistic synthetic datasets with pixel-perfect ground truth annotations. Built on top of the Blender [@blender] 3D creation suite, Syclops offers a flexible and extensible framework for researchers and developers in computer vision, robotics, and related fields. Key features include:
+Syclops is an open-source, modular pipeline for generating large-scale, photorealistic synthetic datasets with pixel-perfect ground truth annotations. Built on Blender’s Cycles engine [@blender], it offers a flexible framework for researchers in computer vision, robotics, and related fields. Key features include:
 
-1. A plugin-based architecture for easy extensibility
-2. Procedural generation of diverse, large-scale environments
-3. Photorealistic rendering using Blender's Cycles engine
-4. Multi-modal sensor simulation (i.e., RGB cameras, depth sensors, stereo depth cameras, structured light)
-5. Accurate ground truth annotation generation
-6. Dynamic scene configuration via YAML-based job files
-7. Scalability to handle environments with millions of objects
+- **Plugin-based architecture** for easy extensibility
+- **Procedural generation** of diverse, large-scale environments
+- **Photorealistic rendering**
+- **Multi-modal sensor simulation** (RGB cameras, depth sensors, stereo cameras)
+- **Comprehensive ground truth annotations**
+- **Dynamic scene configuration** using YAML
+- **Scalability** for millions of objects
 
-Syclops addresses the growing need for high-quality, diverse synthetic data in AI model development and testing, particularly in scenarios where real-world data collection is challenging, expensive, or time-consuming.
+Syclops is especially useful when collecting real-world data is impractical due to cost or difficulty, making it a valuable tool for generating high-quality synthetic data.
 
-# Statement of Need
+## Statement of Need
 
-The performance of machine learning models, especially in computer vision and robotics, heavily depends on the quality and diversity of training data. However, obtaining real-world data can be expensive, time-consuming, and challenging, particularly for rare events or scenarios that are difficult to capture [@tabkhi2022real]. Synthetic data generation offers a promising solution to these challenges, allowing for the creation of large, diverse datasets with accurate ground truth annotations [@mumuni2024survey].
+Machine learning models, particularly in computer vision and robotics, depend on the diversity and quality of training data. Real-world data collection is often expensive and challenging, especially for rare events [@tabkhi2022real]. Synthetic data generation offers an efficient alternative, producing large, annotated datasets [@mumuni2024survey].
 
-Syclops fills an important gap in the landscape of synthetic dataset tools. While several tools exist for generating synthetic data, Syclops distinguishes itself through its focus on large-scale procedural generation, particularly for outdoor and agricultural scenarios. \autoref{tab:comparison} compares Syclops with other prominent synthetic data generation tools.
[…]
