+++ paper.bib
@@ -3,21 +3,27 @@
   author={Weyler, Jan and Magistri, Federico and Marks, Elias and Chong, Yue Linn and Sodano, Matteo and Roggiolani, Gianmarco and Chebrolu, Nived and Stachniss, Cyrill and Behley, Jens},
   journal={IEEE Transactions on Pattern Analysis and Machine Intelligence},
   year={2024},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi={10.1109/TPAMI.2024.3419548}
 }
+
 @inproceedings{greff2022kubric,
   title={Kubric: A scalable dataset generator},
   author={Greff, Klaus and Belletti, Francois and Beyer, Lucas and Doersch, Carl and Du, Yilun and Duckworth, Daniel and Fleet, David J and Gnanapragasam, Dan and Golemo, Florian and Herrmann, Charles and others},
   booktitle={Proceedings of the IEEE/CVF conference on computer vision and pattern recognition},
   pages={3749--3761},
-  year={2022}
+  year={2022},
+  doi={10.1109/CVPR52688.2022.00373}
 }
+
 @article{mumuni2024survey,
   title={A survey of synthetic data augmentation methods in computer vision},
   author={Mumuni, Alhassan and Mumuni, Fuseini and Gerrar, Nana Kobina},
   journal={arXiv preprint arXiv:2403.10075},
-  year={2024}
+  year={2024},
+  doi={10.1007/s11633-022-1411-7}
 }
+
 @article{denninger2023blenderproc2,
   title={Blenderproc2: A procedural pipeline for photorealistic rendering},
   author={Denninger, Maximilian and Winkelbauer, Dominik and Sundermeyer, Martin and Boerdijk, Wout and Knauer, Markus Wendelin and Strobl, Klaus H and Humt, Matthias and Triebel, Rudolph},
@@ -26,27 +32,32 @@
   number={82},
   pages={4901},
   year={2023},
-  publisher={Journal of Open Source Software}
+  publisher={Journal of Open Source Software},
+  doi={10.21105/joss.04901}
 }
+
 @inproceedings{tabkhi2022real,
   title={Real-world computer vision for real-world applications: Challenges and directions},
   author={Tabkhi, Hamed},
   booktitle={Proceedings of SAI Intelligent Systems Conference},
   pages={727--750},
   year={2022},
-  organization={Springer}
+  organization={Springer},
+  doi={10.1007/978-3-031-16072-1_53}
 }
+
 @Manual{blender,
    title = {Blender - a 3D modelling and rendering package},
    author = {Blender Online Community},
    organization = {Blender Foundation},
    address = {Stichting Blender Foundation, Amsterdam},
    year = {2019},
-   url = {http://www.blender.org},
+   url = {http://www.blender.org}
 }
+
 @inproceedings{iqbal2023evaluating,
-  title={Evaluating synthetic vs. real data generation for AI-based selective weeding.},
-  author={Iqbal, Naeem and Bracke, Justus and Elmiger, Anton and Hameed, Hunaid and von Szadkowski, Kai and Hoffmann, Christa and Stein, Anthony and Ruckelshausen, Arno and M{\"u}ller, Henning and Steckel, Thilo},
+  title={Evaluating synthetic vs. real data generation for AI-based selective weeding},
+  author={Iqbal, Naeem and Bracke, Justus and Elmiger, Anton and Hameed, Hunaid and von Szadkowski, Kai and Hoffmann, Christa and Stein, Anthony and Ruckelshausen, Arno and M{"u}ller, Henning and Steckel, Thilo},
   booktitle={GIL Jahrestagung},
   pages={125--135},
   year={2023}
