### 2025-02-26 #7854 @editorialbot
[REVIEW]: Syclops: A Modular Pipeline for Procedural Generation of Synthetic Data
**Submitting author:** @aelmiger (<a href="http://orcid.org/0009-0008-0389-8545">Anton Elmiger</a>)
**Repository:** https://github.com/DFKI-NI/syclops
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v1.4.4
