### 2025-03-13 #7854 @bstanciulescu
generate my checklist
