### 2025-03-25 7985b7c
Dependency Installation Issue - coacd==0.0.5 Not Found Fixes #24
