### 2025-03-25 f7a479b
fix: update integration test dependencies to use correct package names
