### 2025-03-25 314de68
Add References section to paper.md
