### 2025-03-25 aaf5115
docs: add contributing guidelines to the repository
