### 2025-03-25 70b429c
Fix: old coacd pip package (#27)
* Dependency Installation Issue - coacd==0.0.5 Not Found
Fixes #24

* bump version to 1.4.1

* fix: update integration test dependencies to use correct package names

* docs: add contributing guidelines to the repository
