+++ pyproject.toml
@@ -13,7 +13,7 @@ readme = "README.md"
 requires-python = ">=3.8"
 license = {text = "GPLv3"}
 
-version = "1.4.0"
+version = "1.4.1"
 
 dynamic = ["dependencies"]
 
