### 2025-03-28 DFKI-NI/syclops#28 @bstanciulescu
Improve example documentation
This issue is related to the JOSS review: https://github.com/openjournals/joss-reviews/issues/7854
Two main issues arose when trying to test the Syclops installation:
1. The generated example file, is not easy to locate in the folder's tree;
2. Could you detail the way to create its own job files ? The yaml file fields should be explicitly documented.
