### 2025-03-28 #7854 @bstanciulescu
Very impressive work ! The segmentation task , the instance detection and the category detection are amazing !
