### 2025-03-29 DFKI-NI/syclops#29 @sumn2u
pip install . Fails Due to Multiple Top-Level Packages
When attempting to install the `syclops` package using `pip install .`, the build process fails due to setuptools detecting multiple top-level packages, including the virtual environment directory (`syclops_venv`). Additionally, a deprecation warning appears for the `project.license` field in `pyproject.toml`.

#### **Steps to Reproduce**
1. Clone the repository and navigate to the project directory:
   ```sh
   git clone https://github.com/your-repo/syclops.git
   cd syclops
   ```
2. Create a virtual environment inside the project folder:
   ```sh
   virtualenv syclops_venv
   source syclops_venv/bin/activate
   ```
3. Run `pip install .`
4. The following error occurs:
   ```
   error: Multiple top-level packages discovered in a flat-layout: ['syclops', 'syclops_venv'].
   ```

#### **Expected Behavior**
`pip install .` should install the package successfully without detecting the virtual environment as part of the package.

#### **Actual Behavior**
The installation fails because setuptools detects multiple top-level packages.

#### **Error Output (Relevant Logs)**
```sh
(syclops_venv) (base) sk@sumn2u syclops % pip install .
Processing /Users/sumn2u/workspace/syclops
  Installing build dependencies ... done
  Getting requirements to build wheel ... error
  error: subprocess-exited-with-error

  × Getting requirements to build wheel did not run successfully.
  │ exit code: 1
  ╰─> [28 lines of output]
      /private/var/folders/69/wlyjt56j0rd5p015v59m3gn40000gn/T/pip-build-env-a4xzblb4/overlay/lib/python3.11/site-packages/setuptools/config/_apply_pyprojecttoml.py:82: SetuptoolsDeprecationWarning: `project.license` as a TOML table is deprecated
      !!

              ********************************************************************************
              Please use a simple string containing a SPDX expression for `project.license`. You can also use `project.license-files`. (Both options available on setuptools>=77.0.0).

              By 2026-Feb-18, you need to update your project and remove deprecated calls
              or your builds will no longer be supported.

              See https://packaging.python.org/en/latest/guides/writing-pyproject-toml/#license for details.
              ********************************************************************************

      !!
        corresp(dist, value, root_dir)
      error: Multiple top-level packages discovered in a flat-layout: ['syclops', 'syclops_venv'].

      To avoid accidental inclusion of unwanted files or directories,
      setuptools will not proceed with this build.

      If you are trying to create a single distribution with multiple packages
      on purpose, you should not rely on automatic discovery.
      Instead, consider the following options:

      1. set up custom discovery (`find` directive with `include` or `exclude`)
      2. use a `src-layout`
      3. explicitly set `py_modules` or `packages` with a list of names

      To find more information, look for "package discovery" on setuptools docs.
      [end of output]

error: subprocess-exited-with-error

× Getting requirements to build wheel did not run successfully.
│ exit code: 1
╰─> See above for output.

```

#### **Environment**
- Python: 3.11
- `pip`: 25.0.1
- `setuptools`: 75.8.2
- OS: macOS
