### 2025-03-29 #7854 @sumn2u
> [@sumn2u](https://github.com/sumn2u) The issues you created are now addressed and closed.

Thanks, @aelmiger! I appreciate your fixes. I've created another issue and would love your input whenever you have time!
- https://github.com/DFKI-NI/syclops/issues/29
