### 2025-04-02 DFKI-NI/syclops#29 @aelmiger
Thanks for pointing it out. It should be addressed by #30
