### 2025-04-23 356ed28
Bug Report: macOS Not Supported in `install_blender` Fixes #31
