### 2025-04-23 db4f703
Fix: Improve error handling for zip extraction folder renaming
