### 2025-04-23 43c14e0
Bug Report: macOS Not Supported in `install_blender` (#32)
* Bug Report: macOS Not Supported in `install_blender`
Fixes #31

* Fix: Improve error handling for zip extraction folder renaming

* Fix: Bump version to 1.4.3
