### 2025-05-02 DFKI-NI/syclops#33 @sumn2u
FileNotFoundError when running job on macOS
When executing a Syclops job on macOS, the following error occurs after Blender is downloaded but not properly linked:

```sh

(syclops_venv) (base) sumankunwar@Sumans-MacBook-Air syclops % syclops -j syclops/__example_assets__/test_job.syclops.yaml
macOS detected. Please install Blender manually from the downloaded .dmg file.
Downloading Blender for macOS (arm64)...
Downloading blender-3.6.1-macos-arm64.dmg... ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ 100% 0:00:00
Download complete: /Users/sumankunwar/workspace/syclops/blender-3.6.1-macos-arm64.dmg
Please open the .dmg file and drag Blender to your Applications folder.
Syclops folder: /Users/sumankunwar/workspace/syclops
Job Config is valid!
╭─────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────╮
│ ⠋ Waiting for first generated data                                                                                                      │
│ ╭───────────────────────────────────────────────────────────── LAST LOG ──────────────────────────────────────────────────────────────╮ │
│ │ 🌍  Waiting for logs                                                                                                                │ │
│ ╰─────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────╯ │
╰─────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────╯
Traceback (most recent call last):
  File "/Users/sumankunwar/workspace/syclops/syclops_venv/bin/syclops", line 8, in <module>
    sys.exit(main())
             ^^^^^^
  File "/Users/sumankunwar/workspace/syclops/syclops_venv/lib/python3.11/site-packages/syclops/cli.py", line 333, in main
    _run_syclops_job(args, install_folder, Path(args.job_description))
  File "/Users/sumankunwar/workspace/syclops/syclops_venv/lib/python3.11/site-packages/syclops/cli.py", line 285, in _run_syclops_job
    subprocess.run(
  File "/opt/homebrew/Cellar/python@3.11/3.11.9_1/Frameworks/Python.framework/Versions/3.11/lib/python3.11/subprocess.py", line 548, in run
    with Popen(*popenargs, **kwargs) as process:
         ^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/opt/homebrew/Cellar/python@3.11/3.11.9_1/Frameworks/Python.framework/Versions/3.11/lib/python3.11/subprocess.py", line 1026, in __init__
    self._execute_child(args, executable, preexec_fn, close_fds,
  File "/opt/homebrew/Cellar/python@3.11/3.11.9_1/Frameworks/Python.framework/Versions/3.11/lib/python3.11/subprocess.py", line 1955, in _execute_child
    raise child_exception_type(errno_num, err_msg, err_filename)
FileNotFoundError: [Errno 2] No such file or directory: PosixPath('/Users/sumankunwar/workspace/syclops/blender-3.6.1/blender')
```

### ✅ Expected Behavior

Syclops should either:

- Automatically extract and link the Blender binary correctly, **or**
- Clearly instruct the user to install and configure Blender manually *before* attempting to run the job.

---

### 💻 Environment

- **OS:** macOS (arm64)
- **Python:** 3.11.9
- **Syclops version:** 1.4.3

---

### 🔧 Proposed Fix

- Add validation to check whether Blender exists at the expected path before starting the job.
- Provide fallback instructions or an interactive prompt to configure Blender if missing.
