### 2025-05-07 DFKI-NI/syclops#33 @aelmiger
Since we have no access to macOS, we won't be able to support this operating system.
Is it sufficient for you that this is mentioned in the README?
