### 2025-05-07 DFKI-NI/syclops#33 @sumn2u
If you could include this information in the README, that would be great!
By the way, I followed the same setup process on Windows 10 but wasn’t able to get the system running.
I was using Python version 3.11.

Doesn’t this command start the test job application?
`syclops -j syclops/__example_assets__/test_job.syclops.yaml`
