### 2025-05-12 #7854 @boisgera
Hi everyone!

Let me try to summarize the open tasks:

  - @bstanciulescu: could you answer to [this](https://github.com/DFKI-NI/syclops/issues/28#issuecomment-2771632182) so that @aelmiger can address properly the issue https://github.com/DFKI-NI/syclops/issues/28?

 - @sumn2u: AFAICT, @aelmiger fixed https://github.com/DFKI-NI/syclops/issues/29. Is it now ok for you?

- @aelmiger, could you address [this comment](https://github.com/DFKI-NI/syclops/issues/33#issuecomment-2858406983) ?

Please @bstanciulescu and @sumn2u, could you update your review checklists so that we can identify what blocking points may remain? Thank you in advance!

Kind regards,

Sébastien (@boisgera)
