+++ requirements.txt
@@ -6,10 +6,10 @@ filelock==3.9.0
 GitPython==3.1.29
 jsonschema==4.17.3
 opencv-python==4.6.0.66
-numpy==1.23.5
+numpy==1.24.3
 Pillow==9.4.0
 PyYAML==6.0.1
 requests==2.27.1
-ruamel.yaml==0.17.21
+ruamel.yaml==0.18.10
 rich==12.6.0
 wheel==0.38.4
