### 2025-05-13 fab772d
docs: Clarify macOS support status in installation instructions
