### 2025-05-13 DFKI-NI/syclops#33 @aelmiger
Thank you for bringing this to my attention!

I was able to reproduce the issue on Python 3.11. It appears the root cause is an outdated version of a dependency (`ruamel.yaml.clib`). Updating it resolves the issue on my end.

To clarify the usage:

* Run `syclops -c` first to crawl the assets.
* Then use `syclops --example-job` to run the example job.
* Optionally, append `-log` to get more detailed console output for debugging.

Could you try running the following and let me know if it works for you?

```bash
pip install 'ruamel.yaml.clib>=0.2.6'
```

If this resolves the issue on your setup as well, I’ll go ahead and merge the fix.
