### 2025-05-14 DFKI-NI/syclops#33 @sumn2u
I tried using Python `3.9.11` and was able to install the application successfully. However, when I ran the command below, it kept processing without finishing. Do you know how long it typically takes?

```sh
C:\Users\14847\workspace\syclops>syclops -j syclops\__example_assets__\test_job.syclops.yaml
Blender 3.6.1 already installed.
Syclops folder: C:\Users\14847\workspace\syclops
Job Config is valid!
╭───────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────╮
│ ⠋ Waiting for first generated data                                                                                                │
│ ╭────────────────────────────────────────────────────────── LAST LOG ───────────────────────────────────────────────────────────╮ │
│ │ 🌎  2025-05-13 19:36:48,073     INFO root ########################################SCENE                                       │ │
│ │ CONFIGURATION########################################                                                                         │ │
│ │     2025-05-13 19:36:49,736     INFO root Ground: Ground loaded                                                               │ │
│ │     2025-05-13 19:36:49,766     INFO root Environment loaded                                                                  │ │
│ ╰───────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────╯ │
╰───────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────╯
```

I see that the output folder is being generated with the following files:

```sh
PS C:\Users\14847\workspace\syclops\output\2025_05_13_19_25_04> ls

    Directory: C:\Users\14847\workspace\syclops\output\2025_05_13_19_25_04

Mode                 LastWriteTime         Length Name
----                 -------------         ------ ----
d-----         5/13/2025   7:25 PM                tmp
-a----         5/13/2025   7:25 PM           3269 asset_catalog.yaml
-a----         5/13/2025   7:36 PM           7908 blender.log
-a----         5/13/2025   7:25 PM             96 class_id_mapping.yaml
-a----         5/13/2025   7:25 PM           4849 config.yaml
-a----         5/13/2025   7:27 PM            494 logs.log

```
