### 2025-05-14 DFKI-NI/syclops#33 @sumn2u
@aelmiger  Thanks for providing the command. It took me around 5 hours to execute the job, but I was able to get the result. The delay might be due to the testing machine I'm using, which only has 8GB of RAM and uses an 11th Gen Intel(R) Core(TM) i7-1165G7 @ 2.80GHz (2.80GHz), with 4 cores and 8 logical processors.

Here's the final output I obtained.

![Image](https://github.com/user-attachments/assets/b6cb1df0-18cd-4850-916f-3a7b7dc62831)

I suggest mentioning the tested Python version and operating system perhaps in the README or documentation as it provides clearer information about the environment in which the code has been tested.
