### 2025-05-16 fbf0dfb
docs: clarify and update supported Python versions (3.9–3.11)
