### 2025-05-16 eb6fbfb
docs: clarify and update supported Python versions (3.9–3.11) (#35)
