+++ docs/docs/getting_started.md
@@ -5,7 +5,9 @@
 Necessary Software to install on your machine:
 
 
-- Python version 3.9 or higher.
+- ✅ **Tested Python Versions**
+    - Python 3.9 – 3.11
+- ❌ Not yet compatible with Python 3.12+
 
 We recommend using a virtual environment to avoid potential package conflicts. Below are instructions for setting up with `virtualenv` or `conda`.
 
