### 2025-05-17 #7854 @sumn2u
This is an excellent tool for creating synthetic data from 3D virtual environments. The code is well-structured, and I was able to install it successfully using the provided fixes and explore its functionalities. Additionally, the paper is well-written and clearly conveys the significance of this work. Great job, @aelmiger !
