### 2025-05-22 #7854 @boisgera
Hi @bstanciulescu,

Could you please follow-up on

  - https://github.com/DFKI-NI/syclops/issues/28

and (if appropriate) update your checklist?

Thanks in advance!
