### 2025-05-22 DFKI-NI/syclops#28 @bstanciulescu
1. I would suggest that you put it up in the folder tree, as the paths are all relative to the installation folder. If not, a proper indication of its location is sufficient !
2. You are perfectly right, it is my fault. I didn't see the tabs with the fields' description. All good !
I suppose you can close this issue, for my part !
