### 2025-05-22 #7854 @bstanciulescu
> Hi [@bstanciulescu](https://github.com/bstanciulescu),
>
> Could you please follow-up on
>
>     * [Improve example documentation DFKI-NI/syclops#28](https://github.com/DFKI-NI/syclops/issues/28)
>
>
> and (if appropriate) update your checklist?
>
> Thanks in advance!

Done !!
