+++ syclops/postprocessing/postprocessor_base.py
@@ -4,6 +4,7 @@ from pathlib import Path
 from typing import Union
 
 import ruamel.yaml as yaml
+from ruamel.yaml import YAML
 from filelock import FileLock, Timeout
 
 
@@ -122,16 +123,16 @@ class PostprocessorBase(ABC):
             metadata (dict): Dictionary to write
 
         Raises:
-            TimeoutError: If the filelock could not be acquired within 5 seconds.
-        """
+            TimeoutError: If the filelock could not be acquired within 5 seconds.        """
         metadata_file_path = str(
             Path(self.output_folder) / f'{self.config["id"]}_metadata.yaml'
         )
         try:
             lock = FileLock(f"{metadata_file_path}.lock", timeout=5)
             with lock.acquire():
+                yaml_writer = YAML()
                 with open(metadata_file_path, "w") as f:
-                    yaml.dump(metadata, f)
+                    yaml_writer.dump(metadata, f)
         except Timeout:
             error_string = f"Could not acquire filelock for {metadata_file_path}"
             raise TimeoutError(error_string)
