### 2025-06-15 2dcedf3
fix: Adjust numpy version constraint and add importlib_metadata for compatibility
