### 2025-06-16 664cb39
Aelmiger/issue33 (#36)
* Bump version to 1.4.4 and update numpy and ruamel.yaml dependencies

* docs: Clarify macOS support status in installation instructions

* refactor: Update YAML import to use ruamel.yaml consistently across modules

* fix: Adjust numpy version constraint and add importlib_metadata for compatibility

* refactor: Replace ruamel.yaml import style for consistency and improve YAML loading
